Move minSdkVersion to 16, adjust compat infrastructure
