+++ AnkiDroid/src/androidTest/java/com.ichi2.anki.tests/libanki/DBTest.java
@@ -32,13 +32,6 @@ public class DBTest {
     @Test
     public void testDBCorruption() throws Exception {
 
-        // SQLiteDatabase.deleteDatabase() was added in API16.
-        // Corruption handling added in API1 and changed in API11 so implementation's ok
-        // Per Google Play Console 2018/12/11 potential impact is 141 users.
-        if (CompatHelper.getSdkVersion() < Build.VERSION_CODES.JELLY_BEAN) {
-            return;
-        }
-
         String storagePath = CollectionHelper.getDefaultAnkiDroidDirectory();
         File illFatedDBFile = new File(storagePath, "illFatedDB.anki2");
