+++ AnkiDroid/src/main/java/com/ichi2/anki/AnkiActivity.java
@@ -8,10 +8,13 @@ import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.graphics.BitmapFactory;
 import android.graphics.Color;
 import android.media.AudioManager;
 import android.net.Uri;
 import android.os.Bundle;
+
+import androidx.browser.customtabs.CustomTabsIntent;
 import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
@@ -33,6 +36,8 @@ import com.ichi2.anki.dialogs.SimpleMessageDialog;
 import com.ichi2.async.CollectionLoader;
 import com.ichi2.compat.CompatHelper;
 import com.ichi2.compat.customtabs.CustomTabActivityHelper;
+import com.ichi2.compat.customtabs.CustomTabsFallback;
+import com.ichi2.compat.customtabs.CustomTabsHelper;
 import com.ichi2.libanki.Collection;
 import com.ichi2.themes.Themes;
 import com.ichi2.utils.AdaptionUtil;
@@ -307,11 +312,20 @@ public class AnkiActivity extends AppCompatActivity implements SimpleMessageDial
     protected void openUrl(Uri url) {
         //DEFECT: We might want a custom view for the toast, given i8n may make the text too long for some OSes to
         //display the toast
-        if (AdaptionUtil.hasWebBrowser(this)) {
-            CompatHelper.getCompat().openUrl(this, url);
-        } else {
+        if (!AdaptionUtil.hasWebBrowser(this)) {
             UIUtils.showThemedToast(this, getResources().getString(R.string.no_browser_notification) + url, false);
+            return;
         }
+
+        CustomTabActivityHelper helper = getCustomTabActivityHelper();
+        CustomTabsIntent.Builder builder = new CustomTabsIntent.Builder(helper.getSession());
+        builder.setToolbarColor(ContextCompat.getColor(this, R.color.material_light_blue_500)).setShowTitle(true);
+        builder.setStartAnimations(this, R.anim.slide_right_in, R.anim.slide_left_out);
+        builder.setExitAnimations(this, R.anim.slide_left_in, R.anim.slide_right_out);
+        builder.setCloseButtonIcon(BitmapFactory.decodeResource(this.getResources(), R.drawable.ic_arrow_back_white_24dp));
+        CustomTabsIntent customTabsIntent = builder.build();
+        CustomTabsHelper.addKeepAliveExtra(this, customTabsIntent.intent);
+        CustomTabActivityHelper.openCustomTab(this, customTabsIntent, url, new CustomTabsFallback());
     }
 
     public CustomTabActivityHelper getCustomTabActivityHelper() {
