+++ AnkiDroid/src/main/java/com/ichi2/compat/CompatV18.java
@@ -8,6 +8,4 @@ public class CompatV18 extends CompatV17 implements Compat {
 
     @Override
     public long getAvailableBytes(StatFs stat) { return stat.getAvailableBytes(); }
-    @Override
-    public long getTotalBytes(StatFs stat) { return stat.getTotalBytes(); }
 }
